+++ app/src/renderer/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="app">
     <modal-help />
-    <session v-if="config.modals.session.active" />
+    <session v-if="session.modals.session.active" />
     <onboarding v-else-if="onboarding.active" />
     <template v-else>
       <app-header />
@@ -10,10 +10,6 @@
       </div>
     </template>
     <tm-notifications :notifications="notifications" />
-    <modal-error
-      v-if="config.modals.error.active"
-      :body="config.modals.error.message"
-    />
   </div>
 </template>
 
@@ -21,7 +17,6 @@
 import { mapGetters } from "vuex"
 import AppHeader from "common/AppHeader"
 import TmNotifications from "common/TmNotifications"
-import ModalError from "common/TmModalError"
 import ModalHelp from "common/TmModalHelp"
 import Onboarding from "common/TmOnboarding"
 import Session from "common/TmSession"
@@ -33,21 +28,20 @@ import store from "./vuex/store"
  * @vue-prop {Number} [niceProp=1] - A very cool incoerent prop not required but with 1 as default
  * @vue-data {Object} nothing
  * @vue-computed {function} notifications mapGetter
- * @vue-computed {function} config mapGetter
+ * @vue-computed {function} session mapGetter
  * @vue-computed {function} onboarding mapGetter
  */
 export default {
   name: `app`,
   components: {
     AppHeader,
-    ModalError,
     ModalHelp,
     TmNotifications,
     Onboarding,
     Session
   },
   computed: {
-    ...mapGetters([`notifications`, `config`, `onboarding`])
+    ...mapGetters([`notifications`, `session`, `onboarding`])
   },
   mounted() {
     this.$store.commit(`loadOnboarding`)