+++ app/src/renderer/components/common/TmSessionHardware.vue
@@ -57,7 +57,7 @@ export default {
       this.$store.commit(`setModalHelp`, true)
     },
     setState(value) {
-      this.$store.commit(`setModalSessionState`, value)
+      this.$store.commit(`setSessionModalView`, value)
     },
     setStatus(value) {
       this.status = value