+++ app/src/renderer/components/governance/PageProposal.vue
@@ -23,7 +23,7 @@
               <h2 class="page-profile__title">
                 {{ proposal.title }}
               </h2>
-              <h3 v-if="config.devMode">
+              <h3 v-if="session.devMode">
                 Proposer: {{ proposal.proposer || "Fede" }}
               </h3>
             </div>
@@ -181,8 +181,7 @@ export default {
       `connected`,
       `wallet`,
       `votes`,
-      `config`,
-      `user`
+      `session`
     ]),
     proposal() {
       return this.proposals.proposals[this.proposalId]