+++ app/src/renderer/components/staking/PageValidator.vue
@@ -58,7 +58,7 @@
               <dt>My Delegation</dt>
               <dd>{{ myDelegation }}</dd>
             </dl>
-            <dl v-if="config.devMode" class="info_dl colored_dl">
+            <dl v-if="session.devMode" class="info_dl colored_dl">
               <dt>My Rewards</dt>
               <dd>--</dd>
             </dl>
@@ -83,7 +83,7 @@
                 {{ percent(validator.commission.rate) }}
               </dd>
             </dl>
-            <dl v-if="config.devMode" class="info_dl colored_dl">
+            <dl v-if="session.devMode" class="info_dl colored_dl">
               <dt>Slashes</dt>
               <dd>--</dd>
             </dl>
@@ -230,10 +230,9 @@ export default {
       `delegates`,
       `delegation`,
       `committedDelegations`,
-      `config`,
       `keybase`,
       `liquidAtoms`,
-      `user`,
+      `session`,
       `wallet`,
       `connected`
     ]),