+++ test/unit/specs/components/common/TmSessionHardware.spec.js
@@ -38,7 +38,7 @@ describe(`TmSessionHardware`, () => {
       .findAll(`.tm-session-header a`)
       .at(0)
       .trigger(`click`)
-    expect(store.commit.mock.calls[0][0]).toBe(`setModalSessionState`)
+    expect(store.commit.mock.calls[0][0]).toBe(`setSessionModalView`)
     expect(store.commit.mock.calls[0][1]).toBe(`welcome`)
   })